requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
autopep8==2.0.1
boto3==1.26.22
botocore==1.29.22
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.4
defusedxml==0.7.1
dj-database-url==1.0.0
Django==4.1.3
django-activity-stream==1.4.2
django-allauth==0.51.0
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-filter==22.1
django-heroku==0.3.1
django-phonenumber-field==7.0.1
django-phonenumbers==1.0.1
django-seo-js==0.4.1
django-storages==1.13.1
djangorestframework==3.14.0
djangorestframework-oauth==1.1.0
djangorestframework-simplejwt==5.2.2
gunicorn==20.1.0
idna==3.4
jmespath==1.0.1
oauthlib==3.2.2
phonenumbers==8.13.2
Pillow==9.3.0
psycopg2==2.9.5
pycodestyle==2.10.0
pycparser==2.21
PyJWT==2.6.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.7.1
requests==2.28.2
requests-oauthlib==1.3.1
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
tzdata==2022.6
uplink==0.9.7
uritemplate==4.1.1
urllib3==1.26.13
whitenoise==6.2.0
